# packages/dashboard-api/internal/handlers/management_project_cluster.go
package handlers

import (
	"errors"
	"net/http"

	"github.com/gin-gonic/gin"

	"github.com/e2b-dev/infra/packages/dashboard-api/internal/api"
	"github.com/e2b-dev/infra/packages/db/pkg/dberrors"
	"github.com/e2b-dev/infra/packages/shared/pkg/ginutils"
)

func (s *APIStore) ManagementRegisterCluster(c *gin.Context, clusterID api.ClusterID) {
	body, err := ginutils.ParseBody[api.ManagementClusterRegistrationRequest](c.Request.Context(), c)
	if err != nil {
		s.sendAPIStoreError(c, http.StatusBadRequest, "Invalid cluster registration")

		return
	}
	_, err = s.createCluster(c.Request.Context(), clusterRegistration{
		ClusterID:          &clusterID,
		Name:               body.Name,
		Endpoint:           body.Endpoint,
		EndpointTLS:        body.EndpointTls,
		Token:              body.Token,
		SandboxProxyDomain: body.SandboxProxyDomain,
		AuthOrgID:          body.AuthOrgId,
	})
	switch {
	case errors.Is(err, errInvalidClusterRegistration):
		s.sendAPIStoreError(c, http.StatusBadRequest, "Invalid cluster registration")
	case dberrors.IsUniqueConstraintViolation(err), dberrors.IsNotFoundError(err):
		s.sendAPIStoreError(c, http.StatusConflict, "Cluster conflicts with stored state")
	case err != nil:
		s.sendAPIStoreError(c, http.StatusInternalServerError, "Failed to register cluster")
	default:
		c.Status(http.StatusNoContent)
	}
}

func (s *APIStore) ManagementAssignProjectCluster(c *gin.Context, projectID api.ProjectID, clusterID api.ClusterID) {
	s.assignTeamCluster(c, projectID, clusterID, true)
}

func (s *APIStore) ManagementDetachProjectCluster(c *gin.Context, projectID api.ProjectID, clusterID api.ClusterID) {
	s.detachTeamCluster(c, projectID, clusterID)
}

func (s *APIStore) ManagementDeleteCluster(c *gin.Context, clusterID api.ClusterID) {
	s.DeleteAdminClustersClusterID(c, clusterID)
}