# spec/openapi-dashboard.yml
openapi: "3.0.0"
info:
  version: 0.1.0
  title: E2B Dashboard API
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    AdminJWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    # AuthProviderBearerAuth / AuthProviderTeamAuth: B before T in the name
    # so Bearer is validated before Team.
    AuthProviderBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
    AuthProviderTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
  parameters:
...
    clusterID:
      name: clusterID
      in: path
      required: true
      description: Identifier of the cluster.
      schema:
        type: string
        format: uuid
    projectID:
      name: projectID
      in: path
      required: true
      description: Identifier of the project.
      schema:
        type: string
        format: uuid
...
  responses:
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "401":
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "409":
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "500":
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code.
        message:
          type: string
          description: Error message.
...
    ManagementClusterRegistrationRequest:
      type: object
      required:
        - name
        - endpoint
        - endpoint_tls
        - token
      properties:
        name:
          type: string
          minLength: 1
        endpoint:
          type: string
          minLength: 1
        endpoint_tls:
          type: boolean
        token:
          type: string
          minLength: 1
        sandbox_proxy_domain:
          type: string
          nullable: true
        auth_org_id:
          type: string
          nullable: true
...
paths:
...
  /v1/management/clusters/{clusterID}:
    parameters:
      - $ref: "#/components/parameters/clusterID"
    put:
      operationId: managementRegisterCluster
      summary: Register a cluster (v1).
      tags: [control-plane-management]
      security:
        - AdminJWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagementClusterRegistrationRequest"
      responses:
        "204":
          description: Cluster registration is present.
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
    delete:
      operationId: managementDeleteCluster
      summary: Delete an unreferenced cluster (v1).
      description: Deletes a cluster after all team assignments are detached and no active environment references remain. Releases soft-deleted environment references in the same transaction while preserving environment and build history. Repeating a completed deletion succeeds.
      tags: [control-plane-management]
      security:
        - AdminJWTAuth: []
      responses:
        "204":
          description: Cluster is absent.
        "401":
          $ref: "#/components/responses/401"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
  /v1/management/projects/{projectID}/cluster/{clusterID}:
    parameters:
      - $ref: "#/components/parameters/projectID"
      - $ref: "#/components/parameters/clusterID"
    put:
      operationId: managementAssignProjectCluster
      summary: Assign a cluster to a project (v1).
      description: >-
        Assigns the cluster only when the project's tier identifier contains `enterprise`,
        case-insensitively. Replaying the identical assignment succeeds even if the project's
        tier later changes.
      tags: [control-plane-management]
      security:
        - AdminJWTAuth: []
      responses:
        "204":
          description: Cluster is assigned to the project.
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
    delete:
      operationId: managementDetachProjectCluster
      summary: Detach a cluster assignment from a project (v1).
      tags: [control-plane-management]
      security:
        - AdminJWTAuth: []
      responses:
        "204":
          description: The matching assignment is absent.
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"